+++ Example/ManaKit/Set/SetViewModel.swift
@@ -6,7 +6,6 @@
 //  Copyright © 2021 CocoaPods. All rights reserved.
 //
 
-import Foundation
 import Combine
 import CoreData
 import SwiftUI