+++ Package.swift
@@ -5,7 +5,7 @@ import PackageDescription
 
 let package = Package(
     name: "ManaKit",
-    platforms: [.iOS(.v14), .macOS(.v10_15), .tvOS(.v10), .watchOS(.v3)],
+    platforms: [.iOS("15.0"), .macOS(.v10_15), .tvOS(.v14), .watchOS(.v6)],
     products: [
         // Products define the executables and libraries produced by a package, and make them visible to other packages.
         .library(