+++ Sources/Classes/ManaKit+Imaging.swift
@@ -18,15 +18,16 @@ extension ManaKit {
 //        return UIImage(named: imageName.rawValue, in: resourceBundle, compatibleWith: nil)
 //    }
 //
-//    public func symbolImage(name: String) -> UIImage? {
-//        let bundle = Bundle(for: ManaKit.self)
-//        guard let bundleURL = bundle.resourceURL?.appendingPathComponent("ManaKit.bundle"),
-//            let resourceBundle = Bundle(url: bundleURL) else {
-//            return nil
-//        }
-//
-//        return UIImage(named: name, in: resourceBundle, compatibleWith: nil)
-//    }
+    public func symbolImage(name: String) -> UIImage? {
+        let bundle = Bundle(for: ManaKit.self)
+
+        guard let bundleURL = bundle.resourceURL?.appendingPathComponent("ManaKit.bundle"),
+            let resourceBundle = Bundle(url: bundleURL) else {
+            return nil
+        }
+
+        return UIImage(named: name, in: resourceBundle, compatibleWith: nil)
+    }
     
 //    public func downloadImage(ofCard card: MGCard, type: CardImageType, faceOrder: Int) -> Promise<Void> {
 //        guard let url = card.imageURL(type: type,